config/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
geographiclib==2.0
geopy==2.4.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
